requirements.txt
flask
flask-socketio
flask-sqlalchemy
qrcode[pil]
alt-profanity-check
python-dotenv
eventlet
